# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Obras",
    "description": "API de Obras",
    "contact": {
      "name": "Tecnología y Negocios Pelom",
      "url": "https://www.pelom.cl",
      "email": "ti.pelom@pelom.cl"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/mantenedor/v1/creaestado": {
      "post": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "Crea un estado para sae",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nombre": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/mantenedor/v1/creapersona": {
      "post": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "crea una persona",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rut": {
                  "example": "any"
                },
                "apellido_1": {
                  "example": "any"
                },
                "apellido_2": {
                  "example": "any"
                },
                "nombres": {
                  "example": "any"
                },
                "base": {
                  "example": "any"
                },
                "cliente": {
                  "example": "any"
                },
                "id_funcion": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/mantenedor/v1/estados": {
      "get": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "Devuelve todos los estados que puede tomar un reporte",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/mantenedor/v1/findallbases": {
      "get": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "Devuelve todas las bases del sae",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/mantenedor/v1/findallclientes": {
      "get": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "Devuelve todos los clientes sae",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/mantenedor/v1/findalltipofuncionpersonal": {
      "get": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "Devuelve todos los tipo de funcion del personal sae",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/mantenedor/v1/findallpersonas": {
      "get": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "Devuelve todos las personas",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/mantenedor/v1/oficinas": {
      "get": {
        "tags": [
          "SAE - Mantenedores"
        ],
        "description": "Devuelve todas las oficinas del sae",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}